Rename script to BrowserHandler
